api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-docx==1.1.0
PyPDF2==3.0.1
Pillow==10.1.0
pytesseract==0.3.10
groq==0.4.1
python-dotenv==1.0.0
